requirements.txt
flask
flask-socketio
python-socketio
simple-websocket
ddgs
PyPDF2
python-docx
ollama
